# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Yet Another Node.js Blogg Application API",
		"description": "Yet Another Node.js Blogg Application API",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"host": "api-hadministradora.hmg.marlin.com.br",
	"basePath": "/api/v1",
...
	"schemes": [
		"http",
		"https"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
		"/contrato_beneficiario": {
			"post": {
				"tags": [
					"contratos"
				],
				"description": "Create new contratos in system",
				"parameters": [
					{
						"name": "contratos",
						"in": "body",
						"description": "contratos that we want to create",
						"schema": {
							"$ref": "#/definitions/contratos"
						}
					}
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "New contrato is created",
						"schema": {
							"$ref": "#/definitions/contratos"
						}
					}
				}
			},
			"get": {
				"tags": [
					"contratos"
				],
				"parameters": [
					{
						"name": "administradora",
						"in": "query",
						"required": false,
						"description": "administradora",
						"type": "string"
					},
					{
						"name": "operadora",
						"in": "query",
						"required": false,
						"description": "operadora",
						"type": "string"
					},
					{
						"name": "dataNascimento",
						"in": "query",
						"required": false,
						"description": "dataNascimento",
						"type": "string"
					},
					{
						"name": "nomeTitular",
						"in": "query",
						"required": false,
						"description": "nomeTitular",
						"type": "string"
					},
					{
						"name": "entidade",
						"in": "query",
						"required": false,
						"description": "entidade",
						"type": "string"
					},
					{
						"name": "pagina",
						"in": "query",
						"required": false,
						"description": "pagina",
						"type": "string"
					},
					{
						"name": "tamanhoPagina",
						"in": "query",
						"required": false,
						"description": "tamanhoPagina",
						"type": "string"
					}
				],
				"summary": "Get all contratos in system",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/contratos"
						}
					}
				}
			}
		},
...
		"/contrato_beneficiario/csv/{operadoraId}": {
			"parameters": [
				{
					"name": "operadoraId",
					"in": "path",
					"required": true,
					"description": "ID of operadora that we want to create csv",
					"type": "string"
				}
			],
			"get": {
				"tags": [
					"operadora"
				],
				"produces" : [ "application/msexcel" ],
				"summary": "Get operadora with given ID and download csv",
				"responses": {
					"200": {
						"description": "operadora is found",
						
						"schema": {
							"$ref": ""
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{contratoId}": {
			"parameters": [
				{
					"name": "contratoId",
					"in": "path",
					"required": true,
					"description": "ID of contratos that we want to find",
					"type": "string"
				}
			],
			"get": {
				"tags": [
					"contrato"
				],
				"summary": "Get contrato with given ID",
				"responses": {
					"200": {
						"description": "Contrato is found",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			},
			"delete": {
				"summary": "Delete Contrato with given ID",
				"tags": [
					"contrato"
				],
				"responses": {
					"200": {
						"description": "Contrato is deleted",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			},
			"put": {
				"summary": "Update Contrato with give ID",
				"tags": [
					"contrato"
				],
				"parameters": [
					{
						"name": "contrato",
						"in": "body",
						"description": "Contrato with new values of properties",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				],
				"responses": {
					"200": {
						"description": "contrato is updated",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{contratoId}/cobrancas": {
			"parameters": [
				{
					"name": "contratoId",
					"in": "path",
					"required": true,
					"description": "ID of contratos that we want to find",
					"type": "string"
				}
			],
			"post": {
				"tags": [
					"contratos"
				],
				"description": "Create new contratos in system",
				"parameters": [
					{
						"name": "contratos",
						"in": "body",
						"description": "contratos that we want to create",
						"schema": {
							"$ref": "#/definitions/contratos"
						}
					}
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "New contrato is created",
						"schema": {
							"$ref": "#/definitions/contratos"
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{numeroCliente}/titulos": {
			"parameters": [
				{
					"name": "contratoId",
					"in": "path",
					"required": true,
					"description": "ID of contratos that we want to find",
					"type": "string"
				}
			],
			"get": {
				"tags": [
					"contratos"
				],
				"summary": "Get contrato with given ID",
				"responses": {
					"200": {
						"description": "Contrato is found",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{carteirinha}/carteirinha": {
			"parameters": [
				{
					"name": "carteirinha",
					"in": "path",
					"required": true,
					"description": "Numero da Carteirinha",
					"type": "string"
				}
			],
			"get": {
				"tags": [
					"contratos"
				],
				"summary": "Get contrato with given ID",
				"responses": {
					"200": {
						"description": "Contrato is found",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{proposta}/proposta": {
			"parameters": [
				{
					"name": "proposta",
					"in": "path",
					"required": true,
					"description": "Numero Proposta",
					"type": "string"
				}
			],
			"get": {
				"tags": [
					"contratos"
				],
				"summary": "Get contrato with given ID",
				"responses": {
					"200": {
						"description": "Contrato is found",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{cpf}/cpf": {
			"parameters": [
				{
					"name": "cpf",
					"in": "path",
					"required": true,
					"description": "Numero do CPF",
					"type": "string"
				}
			],
			"get": {
				"tags": [
					"contratos"
				],
				"summary": "Get contrato with given ID",
				"responses": {
					"200": {
						"description": "Contrato is found",
						"schema": {
							"$ref": "#/definitions/contrato"
						}
					}
				}
			}
		},
		"/contrato_beneficiario/{idFornecedor}/putFornecedor": {
			"parameters": [
				{
					"name": "IdFornecedor",
					"in": "path",
					"required": true,
					"description": "ID of contratos that we want to find",
					"type": "string"
				}
			],
			
			"put": {
				"summary": "Update Contrato with give ID",
				"tags": [
					"fornecedor"
				],
				"parameters": [
					{
						"name": "fornecedores",
						"in": "body",
						"description": "fornecedor",
						"schema": {
							"$ref": "#/definitions/fornecedor"
						}
					}
				],
				"responses": {
					"200": {
						"description": "contrato is updated",
						"schema": {
							"$ref": "#/definitions/fornecedor"
						}
					}
				}
			}
		}
	},
	"definitions": {
...
		"contrato": {
			"type": "array",
			"$ref": "#/definitons/contratos"
		}
	}
}